Update Go versions on Travis
Both 1.6 and 1.7 have been out for quite a while. I've removed 1.1 as
it's already 4 years old at this point. Unless there is a compelling
reason to keep it I think it's safe to remove.

I've kept Go 1.4 as it was the last version prior to replacing the Go
compiler with one written in Go and as such there are still some people
using it due to it's compilation speed.